# openapi.yaml
openapi: 3.1.0
info:
  title: DeenBridge Backend API
  version: 1.0.0
  license:
    name: MIT
    url: https://github.com/Deen-Bridge/dnb-backend/blob/main/LICENSE
  description: |
    REST API for DeenBridge: authentication and sessions, courses, the book library,
    community spaces, reels, search, Stellar USDC payments, refunds and disputes,
    educator payouts, media upload signing and the SEP-1 metadata document.

    Provenance of this document:
      - Paths, methods and auth requirements are derived from app.js and src/routes/**,
        which are the authoritative mounting and middleware sources.
      - Component schemas are derived from the Mongoose schemas in src/models/**.
      - Per-endpoint request bodies are described from route middleware and the model
        fields they write. Where a controller accepts more than the model exposes, the
        body is intentionally permissive rather than invented.

    Known response inconsistencies (see issue about response standardization):
      - Most handlers return an envelope of the form { success, message, data }.
      - Some list handlers return a bare array with no envelope.
      - POST /api/email returns { success, otp, message } on success but a bare
        { message } on validation failure.
      - GET /admin/jobs returns HTML when the Accept header prefers HTML, otherwise
        { success, jobs }.
      - GET /metrics returns Prometheus text, and the SEP-1 document returns text/toml.
...
servers:
...
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access token issued by POST /api/auth/login or POST /api/auth/refresh.
    jobsToken:
      type: http
      scheme: bearer
      description: Static token from the JOBS_DASHBOARD_TOKEN environment variable.
...
  responses:
    Ok:
      description: Success envelope
      content:
        application/json:
          schema: { $ref: "#/components/schemas/SuccessResponse" }
    Created:
      description: Resource created
      content:
        application/json:
          schema: { $ref: "#/components/schemas/SuccessResponse" }
    BadRequest:
      description: Missing or invalid input
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Unauthorized:
      description: Missing, malformed or expired credentials
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Forbidden:
      description: Authenticated but not permitted for this resource or role
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: Resource does not exist, or the feature is disabled by configuration
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
  schemas:
    ObjectId:
      type: string
      description: MongoDB ObjectId as a 24 character hex string
      pattern: "^[a-fA-F0-9]{24}$"
    Error:
      type: object
      description: Standard error shape produced by the central error handler
      properties:
        success: { type: boolean, examples: [false] }
        error:
          type: object
          properties:
            code:
              type: string
              description: Machine-readable error code
              enum:
                - VALIDATION_ERROR
                - NOT_FOUND
                - UNAUTHORIZED
                - FORBIDDEN
                - CONFLICT
                - RATE_LIMITED
                - PAYLOAD_TOO_LARGE
                - REQUEST_TIMEOUT
                - NETWORK_UNAVAILABLE
                - PAYMENT_FAILED
                - PAYMENT_REJECTED
                - WALLET_MISMATCH
                - WALLET_NOT_CONNECTED
                - ITEM_NOT_FOUND
                - ALREADY_PURCHASED
                - INSUFFICIENT_BALANCE
                - STALE_XDR
                - INVALID_XDR
                - SPONSORSHIP_REJECTED
                - INTERNAL_ERROR
                - DATABASE_ERROR
                - EXTERNAL_SERVICE_ERROR
            message: { type: string, description: Human-readable error message }
            details:
              type: array
              description: Optional field-level validation errors
              items:
                type: object
                properties:
                  field: { type: string }
                  message: { type: string }
        stack: { type: string, description: Present only in development }
        reqId: { type: string, description: Request correlation ID }
    SuccessResponse:
      type: object
      description: Standard success envelope. Some list handlers return a bare array instead.
      properties:
        success: { type: boolean, examples: [true] }
        message: { type: string }
        data: { description: Payload, shape depends on the endpoint }
...
    Review:
      type: object
      properties:
        user: { $ref: "#/components/schemas/ObjectId" }
        comment: { type: string }
        rating: { type: integer, minimum: 1, maximum: 5 }
        createdAt: { type: string, format: date-time }
      required: [user, comment, rating]
...
    Course:
      type: object
      properties:
        _id: { $ref: "#/components/schemas/ObjectId" }
        title: { type: string }
        description: { type: string }
        category: { type: string }
        thumbnail: { type: string, format: uri }
        video: { type: string, description: Video URL or playlist identifier }
        price: { type: number, default: 0, description: Zero means free }
        reviews:
          type: array
          items: { $ref: "#/components/schemas/Review" }
        createdBy: { $ref: "#/components/schemas/ObjectId" }
        enrolledUsers:
          type: array
          items: { $ref: "#/components/schemas/ObjectId" }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
      required: [title, description, category, createdBy]
...
paths:
...
  /api/courses:
    get:
      tags: [Courses]
      summary: List courses
      description: Cached list. Some list handlers in this codebase return a bare array.
      security: []
      responses:
        "200": { $ref: "#/components/responses/Ok" }
      x-envelope-unverified: true
    post:
      tags: [Courses]
      summary: Create a course
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/Course" }
      responses:
        "201": { $ref: "#/components/responses/Created" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /api/courses/user:
    get:
      tags: [Courses]
      summary: List courses by creator
      security: []
      parameters:
        - name: createdBy
          in: query
          required: true
          schema: { $ref: "#/components/schemas/ObjectId" }
      responses:
        "200": { $ref: "#/components/responses/Ok" }
      x-envelope-unverified: true
  /api/courses/recommended:
    post:
      tags: [Courses]
      summary: Recommended courses
      description: Uses POST so a preference payload can be sent. Not cached.
      security: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                interests:
                  type: array
                  items: { type: string }
                userId: { $ref: "#/components/schemas/ObjectId" }
      responses:
        "200": { $ref: "#/components/responses/Ok" }
      x-envelope-unverified: true
  /api/courses/bookmarks:
    get:
      tags: [Courses]
      summary: Bookmarked courses for the caller
      responses:
        "200": { $ref: "#/components/responses/Ok" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /api/courses/{id}/bookmark:
    post:
      tags: [Courses]
      summary: Toggle a course bookmark
      parameters:
        - name: id
          in: path
          required: true
          schema: { $ref: "#/components/schemas/ObjectId" }
      responses:
        "200": { $ref: "#/components/responses/Ok" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      tags: [Courses]
      summary: Remove a course bookmark
      parameters:
        - name: id
          in: path
          required: true
          schema: { $ref: "#/components/schemas/ObjectId" }
      responses:
        "200": { $ref: "#/components/responses/Ok" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /api/courses/{id}/bookmark/check:
    get:
      tags: [Courses]
      summary: Whether a course is bookmarked
      parameters:
        - name: id
          in: path
          required: true
          schema: { $ref: "#/components/schemas/ObjectId" }
      responses:
        "200": { $ref: "#/components/responses/Ok" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /api/courses/{id}:
    get:
      tags: [Courses]
      summary: Get a course
      security: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { $ref: "#/components/schemas/ObjectId" }
      responses:
        "200": { $ref: "#/components/responses/Ok" }
        "404": { $ref: "#/components/responses/NotFound" }
    put:
      tags: [Courses]
      summary: Update a course
      parameters:
        - name: id
          in: path
          required: true
          schema: { $ref: "#/components/schemas/ObjectId" }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/Course" }
      responses:
        "200": { $ref: "#/components/responses/Ok" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/courses/{id}/enroll:
    post:
      tags: [Courses]
      summary: Enrol the caller in a course
      parameters:
        - name: id
          in: path
          required: true
          schema: { $ref: "#/components/schemas/ObjectId" }
      responses:
        "200": { $ref: "#/components/responses/Ok" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/courses/{id}/reviews:
    post:
      tags: [Courses]
      summary: Add a course review
      parameters:
        - name: id
          in: path
          required: true
          schema: { $ref: "#/components/schemas/ObjectId" }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                comment: { type: string }
                rating: { type: integer, minimum: 1, maximum: 5 }
              required: [comment, rating]
      responses:
        "201": { $ref: "#/components/responses/Created" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/courses/{id}/certificate:
    post:
      tags: [Courses, Certificates]
      summary: Generate verifiable course completion certificate
      parameters:
        - name: id
          in: path
          required: true
          schema: { $ref: "#/components/schemas/ObjectId" }
      responses:
        "201": { $ref: "#/components/responses/Created" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
...
  /api/books:
    get:
      tags: [Books]
      summary: List books
      security: []
      responses:
        "200": { $ref: "#/components/responses/Ok" }
      x-envelope-unverified: true
    post:
      tags: [Books]
      summary: Create a book with cover and file upload
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                thumbnail: { type: string, format: binary }
                file: { type: string, format: binary }
                title: { type: string }
                description: { type: string }
                category: { type: string }
                price: { type: number }
              required: [title, description]
      responses:
        "201": { $ref: "#/components/responses/Created" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
...
  /api/books/recom:
    get:
      tags: [Books]
      summary: Fetch recommended books by user interests or popular fallback
      security: []
      parameters:
        - name: interests
          in: query
          required: false
          schema: { type: string }
      responses:
        "200": { $ref: "#/components/responses/Ok" }
        "400": { $ref: "#/components/responses/BadRequest" }
    post:
      tags: [Books]
      summary: Fetch recommended books by user interests payload
      security: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                interests:
                  oneOf:
                    - type: array
                      items: { type: string }
                    - type: string
      responses:
        "200": { $ref: "#/components/responses/Ok" }
        "400": { $ref: "#/components/responses/BadRequest" }
  /api/books/bookmarks:
    get:
      tags: [Books]
      summary: Bookmarked books for the caller
      responses:
        "200": { $ref: "#/components/responses/Ok" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /api/books/{id}/bookmark:
    post:
      tags: [Books]
      summary: Toggle a book bookmark
      parameters:
        - name: id
          in: path
          required: true
          schema: { $ref: "#/components/schemas/ObjectId" }
      responses:
        "200": { $ref: "#/components/responses/Ok" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      tags: [Books]
      summary: Remove a book bookmark
      parameters:
        - name: id
          in: path
          required: true
          schema: { $ref: "#/components/schemas/ObjectId" }
      responses:
        "200": { $ref: "#/components/responses/Ok" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /api/books/{id}/bookmark/check:
    get:
      tags: [Books]
      summary: Whether a book is bookmarked
      parameters:
        - name: id
          in: path
          required: true
          schema: { $ref: "#/components/schemas/ObjectId" }
      responses:
        "200": { $ref: "#/components/responses/Ok" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /api/books/by-author/{authorId}:
    get:
      tags: [Books]
      summary: List books by author
      security: []
      parameters:
        - name: authorId
          in: path
          required: true
          schema: { $ref: "#/components/schemas/ObjectId" }
      responses:
        "200": { $ref: "#/components/responses/Ok" }
      x-envelope-unverified: true
  /api/books/{id}/preview:
    get:
      tags: [Books]
      summary: Stream a limited book preview
      parameters:
        - name: id
          in: path
          required: true
          schema: { $ref: "#/components/schemas/ObjectId" }
      responses:
        "200":
          description: Streamed file bytes
          content:
            application/pdf:
              schema: { type: string, format: binary }
            application/octet-stream:
              schema: { type: string, format: binary }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/books/{id}:
    get:
      tags: [Books]
      summary: Get a book
      security: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { $ref: "#/components/schemas/ObjectId" }
      responses:
        "200": { $ref: "#/components/responses/Ok" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      tags: [Books]
      summary: Delete a book
      parameters:
        - name: id
          in: path
          required: true
          schema: { $ref: "#/components/schemas/ObjectId" }
      responses:
        "200": { $ref: "#/components/responses/Ok" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/books/{id}/reviews:
    post:
      tags: [Books]
      summary: Add a book review
      parameters:
        - name: id
          in: path
          required: true
          schema: { $ref: "#/components/schemas/ObjectId" }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                comment: { type: string }
                rating: { type: integer, minimum: 1, maximum: 5 }
              required: [comment, rating]
      responses:
        "201": { $ref: "#/components/responses/Created" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
...
  /api/purchase/book:
    post:
      tags: [Books]
      summary: Record a book purchase for the caller
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                bookId: { $ref: "#/components/schemas/ObjectId" }
              required: [bookId]
      responses:
        "200": { $ref: "#/components/responses/Ok" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
...